Flyway demonstration
Rather than letting Hibernate create the schema for us, we add flyway
with an initial script that creates the schema and populates it.

As Flyway has been detected, Spring Boot no longer requires Hibernate to
create the schema for us on an embedded database. The result of the
migration can be checked using the `/actuator/flyway` endpoint.