Add local profile to connect to another database
This commit adds an additional configuration file to connect to a
different database. For this demo, we use MySQL. If the "local" profile
is enabled, the in-memory database is no longer created and we connect
to our local persistent database instead.

Since we have flyway, the database is initialized transparently for us.